Add SRM segmentation benchmark script

+++ segmentation/srm/srm.sh
@@ -1,8 +1,14 @@
 #!/bin/bash
 
-MATLAB_HOME=~/MATLAB/R2013b/
+if [ $# -lt 2 ]; then
+    echo "No arguments. Please provide path to MATLAB home directory and images directory"
+    exit 1
+fi
+
+MATLAB_HOME=$1
 MATLAB=$MATLAB_HOME/bin/matlab
 MEX=$MATLAB_HOME/bin/mex
+IMAGES_DIR=$2
 
 rm -Rf out
 mkdir out
@@ -11,5 +17,5 @@ echo "compiling mex"
 $MEX srm_boundarygradient.c
 $MEX srm_boundarylen.c
 
-echo "running demo"
-$MATLAB -nodisplay -r "srm_experiment('../../data/validation/')"
+echo "running experiment"
+$MATLAB -nodisplay -r "srm_experiment('$IMAGES_DIR')"
